AUTH-5 refactor
